+++ src/main/java/com/example/donttouchme/event/domain/Tag.java
@@ -1,8 +1,6 @@
-package com.example.donttouchme.tag.domain;
+package com.example.donttouchme.event.domain;
 
 import com.example.donttouchme.common.Entity.BaseEntity;
-import com.example.donttouchme.event.domain.Event;
-import com.example.donttouchme.eventdetail.domain.TagEventDetail;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
