+++ src/main/java/com/example/donttouchme/event/domain/TagEventDetail.java
@@ -1,7 +1,7 @@
-package com.example.donttouchme.eventdetail.domain;
+package com.example.donttouchme.event.domain;
 
 import com.example.donttouchme.common.Entity.BaseEntity;
-import com.example.donttouchme.tag.domain.Tag;
+import com.example.donttouchme.eventdetail.domain.EventDetail;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
