+++ src/main/java/com/example/donttouchme/event/repository/TagRepository.java
@@ -1,6 +1,6 @@
-package com.example.donttouchme.tag.repository;
+package com.example.donttouchme.event.repository;
 
-import com.example.donttouchme.tag.domain.Tag;
+import com.example.donttouchme.event.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
