+++ src/test/java/com/example/donttouchme/support/IntegrationTestSupport.java
@@ -8,7 +8,7 @@ import com.example.donttouchme.eventdetail.domain.EventDetail;
 import com.example.donttouchme.member.domain.Member;
 import com.example.donttouchme.member.domain.value.LoginProvider;
 import com.example.donttouchme.member.domain.value.ROLE;
-import com.example.donttouchme.tag.domain.Tag;
+import com.example.donttouchme.event.domain.Tag;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
@@ -28,6 +28,17 @@ public abstract class IntegrationTestSupport {
                 .loginProvider(LoginProvider.original)
                 .contact("010-1234-5678")
                 .password("test")
+                .builderWithPassword();
+    }
+
+    protected Member createTestMemberForEvent() {
+        return Member.builderWithPassword()
+                .name("testMember")
+                .email("testMember@test.com")
+                .role(ROLE.user)
+                .loginProvider(LoginProvider.original)
+                .contact("010-1234-5678")
+                .password("test123131231")
                 .builderWithPassword();
     }
